move tool to `bundle` and reuse `bin_parse`
