+++ bundle/Cargo.toml
@@ -3,8 +3,14 @@ name = "bundle"
 version = "0.0.0"
 edition = "2021"
 
+[[bin]]
+name = "proto-bin-to-json"
+path = "src/main.rs"
+required-features = ["build-binary"]
+
 [dependencies]
 anyhow = "1.0.44"
+clap = { version = "4.4.18", features = ["derive", "env"], optional = true }
 codeowners = { path = "../codeowners" }
 chrono = { version = "0.4.38", features = ["serde"] }
 proto = { path = "../proto" }
@@ -65,3 +71,4 @@ pyo3 = [
   "codeowners/pyo3",
   "context/pyo3",
 ]
+build-binary = ["clap"]
