+++ proto/Cargo.toml
@@ -4,21 +4,11 @@ edition = "2021"
 version = "0.0.0"
 build = "src/build.rs"
 
-[[bin]]
-name = "proto-bin-parser"
-path = "src/main.rs"
-required-features = ["build-binary"]
-
 [lib]
 name = "proto"
 path = "src/lib.rs"
 
-[features]
-build-binary = ["anyhow", "clap"]
-
 [dependencies]
-anyhow = { version = "1.0.44", optional = true }
-clap = { version = "4.4.18", features = ["derive", "env"], optional = true }
 prost = "0.12.6"
 prost-types = "0.12.6"
 prost-wkt-types = { version = "0.5.1", features = ["vendored-protox"] }
